+++ activerecord/lib/active_record/connection_adapters/sqlite3_adapter.rb
@@ -102,6 +102,7 @@ module ActiveRecord
         binary:       { name: "blob" },
         boolean:      { name: "boolean" },
         json:         { name: "json" },
+        enum:         { name: "varchar" },
       }
 
       DEFAULT_PRAGMAS = {